+++ pyiron_base/storage/has_stored_traits.py
@@ -75,7 +75,7 @@ class HasStoredTraits(HasTraits, HasStorage, ABC, metaclass=ABCTraitsMeta):
         ... )
         >>>
         >>> from pyiron_base.interfaces.has_hdf import HasHDF
-        >>> from pyiron_base.storage.input import HasStoredTraits
+        >>> from pyiron_base.storage.has_stored_traits import HasStoredTraits
         >>>
         >>>
         >>> class Omelette(HasStoredTraits):
